% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{deg_count}
\alias{deg_count}
\title{Get degree counts
Returns the degree counts from a degree sequence}
\usage{
deg_count(degs)
}
\arguments{
\item{degs}{integer vector of degrees}
}
\description{
Get degree counts
Returns the degree counts from a degree sequence
}
