% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{deg_surv}
\alias{deg_surv}
\title{Get degree survival}
\usage{
deg_surv(degs)
}
\arguments{
\item{degs}{integer vector of degrees}
}
\description{
Returns empirical survival from degree sequence as a data frame
}
