% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dists.R
\name{dppa}
\alias{dppa}
\title{PMF of PPA model}
\usage{
dppa(x, n0, a, b, eps, lambda, log = T)
}
\arguments{
\item{x}{Numeric or numeric vector of values for the PMF to be evaluated at}

\item{n0}{Value of n0 in PA function}

\item{a}{Value of a in PA function}

\item{b}{Value of b in PA function}

\item{eps}{Value of eps in PA function}

\item{lambda}{Value of lambda usually found using find_lambda}
}
\description{
Returns log of PMF
}
\seealso{
Other distribution functions: 
\code{\link{pppa}()}
}
\concept{distribution functions}
