% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_funcs.R
\name{find_lambda}
\alias{find_lambda}
\title{Finding the value of \code{lambda}}
\usage{
find_lambda(n0, a, b, eps)
}
\description{
Function to be used to find the value of the constant \code{lambda} in the Rudas formula for a set of parameters.
}
