% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample_ppa}
\alias{sample_ppa}
\title{Sampling from PPA model}
\usage{
sample_ppa(N, a, b, n0, smplr = NULL, ..., init_size = 25, quiet = T)
}
\arguments{
\item{N}{Number of network generation steps to run.}

\item{a}{Value of a in PA function}

\item{b}{Value of b in PA function}

\item{n0}{Value of n0 in PA function}

\item{smplr}{function to be used to sample edges (default is NULL and will sample a tree)}

\item{...}{parameters to be passes to \code{smplr}}

\item{init_size}{initial size of the network (only used when \code{smplr}!=NULL)}

\item{quiet}{boolean}
}
\description{
Sampling from PPA model
}
