\name{twbfn}
\alias{twbfn}
\alias{twbfn}
\docType{package}
\title{
  Collection of functions for Thomas Boughen's PhD
}
\description{
  This R package has been made in order to collect the functions that I have been using for sampling from certain network models and investigating their properties.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Thomas W Boughen,<t.w.boughen1@newcastle.ac.uk>.

Maintainer: Thomas W Boughen <t.w.boughen1@newcastle.ac.uk>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
