% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat.R
\name{xi_heatmap}
\alias{xi_heatmap}
\title{Make heatmap for value of xi}
\usage{
xi_heatmap(
  n0 = 15,
  focus_xi = 0.5,
  a_lims = c(0, 2),
  b_lims = c(0, 2),
  res = 200
)
}
\description{
Make heatmap for value of xi
}
