% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{ppa_mcmc}
\alias{ppa_mcmc}
\title{MCMC Sampling for PPA Model}
\usage{
ppa_mcmc(
  data,
  init_n0,
  init_a,
  init_b,
  init_lambda,
  n_iter = 10000,
  a_sd = 0.1,
  b_sd = 0.1,
  n0_sd = 1,
  adapt_interval = 1000,
  burn_in = 10000
)
}
\arguments{
\item{data}{A numeric matrix containing the data, with the first column being the unique degrees and the second being the
counts of those degrees.}

\item{init_n0}{Initial value for the parameter \code{n0}.}

\item{init_a}{Initial value for the parameter \code{a}.}

\item{init_b}{Initial value for the parameter \code{b}.}

\item{init_lambda}{Initial value for the parameter \code{lambda}.}

\item{n_iter}{Number of iterations (default is 10000).s}

\item{a_sd}{Initial standard deviation for \code{a} proposals (default is 0.1).}

\item{b_sd}{Initial standard deviation for \code{b} proposals (default is 0.1).}

\item{n0_sd}{Initial standard deviation for \code{n0} proposals (default is 1).}

\item{adapt_interval}{Interval for adapting the proposal variance (default is 1000).}

\item{burn_in}{Period where MCMC is not adaptive.}
}
\value{
A list containing the results.
}
\description{
This function attempts to estimate the parameters of the preferential attachment function
from the sequence of degrees of a network given that it was generated using the model
from Rudas.
}
\seealso{
Other model fitting functions: 
\code{\link{ppa_mle}()}
}
\concept{model fitting functions}
